run_tests.py: fix cleanup of temporary directories after the test.
Previously a successful test run left behind some temporary directories.
Now all such directories are removed after the test. If some tests fail,
then the corresponding directories and the files in them are left
untouched (to allow investigation of the failed tests). This corresponds
to the behaviour of the old bash testing script run_tests.sh.